requirements.txt
aiogram
python-dotenv
langchain
langchain_community
transformers